move custom drivers, enable SPI flag, and add HW_reg_acces.S
