+++ .config
@@ -106,7 +106,7 @@ CONFIG_SERVICE_QSPI_MICRON_MQ25T=y
 CONFIG_SERVICE_REBOOT=y
 # CONFIG_SERVICE_SCRUB is not set
 CONFIG_SERVICE_SGDMA=y
-# CONFIG_SERVICE_SPI is not set
+CONFIG_SERVICE_SPI=y
 CONFIG_SERVICE_TINYCLI=y
 
 #
