+++ baremetal/polarfire-soc-bare-metal-library/src/platform/drivers/mss/CoreSPI/corespi_regs.h
