+++ config.h
@@ -2,6 +2,7 @@
 #define CONFIG_DEFAULT_DEVICE_TREE "boards/mpfs-icicle-kit-es/mpfs_minimal_uboot.dtb"
 #define CONFIG_SERVICE_BEU 1
 #define CONFIG_SERVICE_BOOT 1
+#define CONFIG_SERVICE_BOOT_SPI_FLASH_OFFSET 0x400
 #define CONFIG_SERVICE_BOOT_DDR_TARGET_ADDR 0x103FC00000
 #define CONFIG_SERVICE_DDR 1
 #define CONFIG_SERVICE_GOTO 1
@@ -39,6 +40,7 @@
 #define CONFIG_SERVICE_QSPI_MICRON_MQ25T 1
 #define CONFIG_SERVICE_REBOOT 1
 #define CONFIG_SERVICE_SGDMA 1
+#define CONFIG_SERVICE_SPI 1
 #define CONFIG_SERVICE_TINYCLI 1
 #define CONFIG_SERVICE_TINYCLI_TIMEOUT 1
 #define CONFIG_SERVICE_USBDMSC 1
